Refactored standalone modules into `utils` package
The issue was Sound Player being in the `app/lib` package meant it was required to load all of the GUI to use it.  This was a problem when trying to run Scan Manager outside of the GUI ecosystem.  This organization fixes this and other similar issues, and in the future we'll try to put modules with no dependence on any `app` or `data_acquisition` modules into this `utils` package.
