+++ GEECS-Scanner-GUI/geecs_scanner/app/lib/__init__.py
@@ -1,4 +1,2 @@
 from .menu_bar_option import MenuBarOption, MenuBarOptionBool, MenuBarOptionStr
 from .action_control import ActionControl
-from . import module_open_folder
-from .sound_player import SimpleSoundPlayer, SoundPlayer, multiscan_finish_jingle
