+++ GEECS-Scanner-GUI/geecs_scanner/utils/application_paths.py
@@ -1,21 +1,5 @@
-import sys
-import traceback
 import configparser
 from pathlib import Path
-
-
-def exception_hook(exctype, value, tb):
-    """
-    Global wrapper to print out tracebacks of python errors during the execution of a PyQT window.
-
-    :param exctype: Exception Type
-    :param value: Value of the exception
-    :param tb: Traceback
-    """
-    print("An error occurred:")
-    traceback.print_exception(exctype, value, tb)
-    sys.__excepthook__(exctype, value, tb)
-    sys.exit(1)
 
 
 class ApplicationPaths:
@@ -24,7 +8,7 @@ class ApplicationPaths:
     """
 
     CONFIG_PATH = Path('~/.config/geecs_python_api/config.ini').expanduser()
-    BASE_PATH = Path(__file__).parents[1] / "scanner_configs" / "experiments"
+    BASE_PATH = Path(__file__).parents[2] / "scanner_configs" / "experiments"
     SAVE_DEVICES_FOLDER = "save_devices"
     SCAN_DEVICES_FOLDER = "scan_devices"
     PRESET_FOLDER = "scan_presets"
