+++ GEECS-Scanner-GUI/geecs_scanner/utils/module_open_folder.py
